requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
python-multipart>=0.0.5,<0.1.0
pymongo>=4.0.0,<5.0.0
pandas>=1.3.0,<2.0.0
python-dotenv>=0.19.0,<0.20.0
transformers>=4.30.0,<5.0.0
torch>=2.0.0,<3.0.0
sentence-transformers>=2.2.0,<3.0.0
streamlit>=1.22.0,<2.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0